+++ plone/app/event/interfaces.py
@@ -5,6 +5,9 @@ from plone.event.utils import default_timezone as fallback_default_timezone
 from plone.app.event import messageFactory as _
 
 
+ISO_DATE_FORMAT = '%Y-%m-%d'
+
+
 class IEvent(Interface):
     """ Generic Event
     """
@@ -31,14 +34,17 @@ class IOccurrence(Interface):
     start = Attribute(u"Occurrence start date")
     end = Attribute(u"Occurrence end date")
 
-## Adapter interfaces
 
+## Adapter interfaces
 class IRecurrence(Interface):
     """ Adapter for recurring events.
     """
     def occurrences(limit_start, limit_end):
-        """ Return the occurrences of the recurring event.
         """
+        Return the occurrences of the recurring event. Each
+        occurrence is providing IOccurrence.
+        """
+
 
 class IICalendar(Interface):
     """ Adapter, which is used to construct an icalendar object.